services/tournaments/requirements.txt
Django>=4.1,<4.2
djangorestframework
djangorestframework-simplejwt
redis==4.6.0 